Tech Debt - Remove duplicate AWSError classes
