+++ moto/sagemaker/exceptions.py
@@ -1,5 +1,4 @@
 from __future__ import unicode_literals
-import json
 from moto.core.exceptions import RESTError, JsonRESTError
 
 ERROR_WITH_MODEL_NAME = """{% extends 'single_error' %}
@@ -27,22 +26,6 @@ class MissingModel(ModelError):
     def __init__(self, *args, **kwargs):
         super(MissingModel, self).__init__(
             "NoSuchModel", "Could not find model", *args, **kwargs
-        )
-
-
-class AWSError(Exception):
-    TYPE = None
-    STATUS = 400
-
-    def __init__(self, message, type=None, status=None):
-        self.message = message
-        self.type = type if type is not None else self.TYPE
-        self.status = status if status is not None else self.STATUS
-
-    def response(self):
-        return (
-            json.dumps({"__type": self.type, "message": self.message}),
-            dict(status=self.status),
         )
