+++ moto/sagemaker/responses.py
@@ -2,9 +2,9 @@ from __future__ import unicode_literals
 
 import json
 
+from moto.core.exceptions import AWSError
 from moto.core.responses import BaseResponse
 from moto.core.utils import amzn_request_id
-from .exceptions import AWSError
 from .models import sagemaker_backends
