+++ moto/xray/responses.py
@@ -3,10 +3,11 @@ import json
 import datetime
 
 from moto.core.responses import BaseResponse
+from moto.core.exceptions import AWSError
 from six.moves.urllib.parse import urlsplit
 
 from .models import xray_backends
-from .exceptions import AWSError, BadSegmentException
+from .exceptions import BadSegmentException
 
 
 class XRayResponse(BaseResponse):
